[Kernels] Use `expm1` inplace of `exp(x)-1` in the Elu activation (#16258)
Replace the use of `exp(x)-1` with `expm1(x)` for the elu activation.
Note that this will call into libm, but that's fine for now. We will
revisit if we have to have an optimal expm1 in a followup pr.

NN_ORIG_REV_ID: 106a80ef5ac7ae0a7d46dc37ac5151cb384f73de