[MOGG] Add param to tell the matmul when we have fused (#16702)
The matmul is able to pick a slightly faster path when the lambda is an
identity lambda. We don't actually fuse right now but we have tests to
check it works. When we fuse we will set this flag to true and the
matmul will use the runtime closure instead

MOGG_ORIG_REV_ID: c4c4e456fabf8cd1e888fd6189a36b5c614cb2eb