[Kernels] Unify duplicate gather kernels (#17815)
Previously a duplicate gather kernel was added to MOGG because the
existing one did not support input/output fusion and the axis was
expected to be a parameter. This unifies the gather kernels by extending
the the input/output fusion gather generator and using it as a common
base. The new gather generator supports a prefetch function hook, and
makes the axis an OptionalParamInt so that the shim (used for
benchmarking) can fill in the axis as a parameter, and MOGG can fill in
the axis as a runtime value, which gets inlined by the jit during const
propagation.

MOGG_ORIG_REV_ID: acef1f99c08a9226c04bdd8c0845407836187893