[Kernels] Use memcpy-based concat in certain scenarios (#18271)
[Internal link] removed the memcpy-based
concat kernel because a simple concat was better for a tiny model, which
caused a large regression in DLRM which has a big concat.

Also removes uses of _NDBufferVector in concat because it is no longer
needed and requires a bunch of copying from the VariadicList to the
vector.

Fixes [Internal link]

MOGG_ORIG_REV_ID: c65d6b56115e38cb145d7eb381d3313fa054029f