[Kernels] cleanup matmul load/store c tile (#20174)
Cleanup change in advance of deeper load/store tile work: The existing
code would do bounds checking for the number of rows to load/store, but
this is not needed as the number of C rows is known to already exist.
There are kernel designs where the number of rows to output may be less
than the internal microkernel tile, but that is not the case here.

Removing this check saves a test at each vector of the
skip_col_bounds=True generated kernels. If this assertion is not true,
then that messes up the other work I was doing. :)

This also removes the aligned load/stores in this code as this adds no
value to any modern vector implementation.

LINALG_ORIG_REV_ID: d40092eb494ebb84da7ee446b961ad7f820b801b