[Kernels] optimize inner loop of NEON matmul (#20229)
This optimizes the inner loop used for the NEON matmul kernel.
Performance benefits described below or in comments.

The NEON matmul currently uses two kernel paths: one for all GEMMs where
K is a multiple of 4096 (reference #13814) and everything else. From
digging into this more, the problem that Graviton2 had was doing a read
of A immediately before the use of that vector. If the read of A was
"slow", such as a read to an aligned cache offset from the last read,
then the inner loop of the matmul would stall and performance would
suffer. Another observation was that using the latest LLVM that had a
Graviton2 CPU schedule, the performance of the original kernel was not
as bad because the reads of A were scheduled much earlier in the loop to
avoid the chance for stalls.

From measuring the ARM Compute Library with the shapes, that kernel also
did not need to handle the multiple of 4096 case specially. The biggest
difference from that kernel and the way our NEON matmul kernel behaved
is that all of the loads from A are scheduled very early while the loads
from B tend to be more streaming like and are distributed through the
inner kernel.

So this change adapts that pattern and sees a good jump in performance
for Graviton3 and M1 systems. Graviton2 does seen an increase for the K
multiple of 4096 case, but the other cases are roughly the same.

This change does not yet remove the split between K multiple of 4096
path and the normal path. These two paths still use a different inner
kernel size that can be addressed in a follow on PR. But these paths are
now converged and use the same inner kernel template. There is a code
size increase from this because the inner loop is larger from the K
multiple of 4096 path, but totally worth it.

The use of `@adaptive` in the _accumulate overloads has always been
confusing to me and I changed the NEON _run_inner_loop to call a
different name function instead. I purposely wanted the scalar tail case
to use the same template as the vector body case and that didn't work
with the old scheme.

LINALG_ORIG_REV_ID: c8bf6a7567081781aafd6059e1a878b8d764989f