[Stdlib] Have a singleton default LLCL Runtime (#21926)
The normal construction of runtime is without args. I.e.

```
with Runtime() as rt:
  something_using_rt(rt, ...)
```

This makes that usage utilize a single global singleton runtime that is
constructed once. With this in place, we can depricate the parallelize
that takes a runtime -- which was added for perf reasons.

Part of [Internal link]

RUNTIME_ORIG_REV_ID: f439ed0f8b50874f4e87ebc63d00d1563eaf3cb0