[Stdlib] Remove the parallelize function which takes a runtime arg (#21942)
Now that we have a singleton runtime (introduced in #21926), we can lean
more into that and remove the parallelize runtime. There were a few
examples that used the parallelize with a global runtime, these are no
longer necessary so remove them.

Also, this adds a vague improvement message to the changelong. The Mojo
user has no idea about the LLCL runtime and we never documented it. From
their perspective this is black magic, and we should probably keep it
this way until we commit to documenting and explaining LLCL.

Closes [Internal link]

ALGORITHM_ORIG_REV_ID: 27ac0e6f39d853211e1b2b464251f068dc56c33d