[MOGG] Deduce parameters using lambdas (#21789)
Related  #21380 (fix will come in another PR)

Using lambdas in the IR to deduce the parameters makes it possible to
get rid of the `NDBuffer` outputs from most kernels (except matmuls and
convolutions because they are used by the kernels which is a breach of
the lambda contract). It also simplifies the lambda signature that we
pass. Instead of having a data type parameter in the lambda, we make it
inherit the parameter of the kernel itself which is what we do in the
body anyway.

MOGG_ORIG_REV_ID: 4cb408af7c493c875374cc6451a8ed89b9a0cfa4