[Stdlib][Kernels] Split up the tests into different targets/dirs, NFC (#22429)
This splits up the tests to make the tests more granular and enable
faster tests and iteration time when making modifications. The idea is
to mirror the dir structure already in the the mojo dir.

Part of #21981

MATH_ORIG_REV_ID: bd5734e6a03c0c29b807c2256810a382fcebaf80