[KGEN] Have add_mojo_package understand nested files (#23353)
The GPU package is nested and add_mojo_package did not work with input
files with slashes. This fixes that so that we can track the deps and
correctly rebuild the package if a file within it changes.

Closes #23345

GPU_ORIG_REV_ID: 04a18a62509e7d9d4af0e8d985cb9abb05f443b8