[Kernels] Introduce parallel_stencil generator (#23382)
Parallel stencil generator is similar to a variant of elementwise
parallel generator, for each point in the output domain it defines a continuous range
(stencil hyperrectangle) where a function
is computed and the stencil is reduced to a single point.

This generator can implement various operations including Pooling,
depthwise conv that is fused with padding.

In this PR:
  Adding functional.parallel_stencil
  Add tests for implementing pooling, pooling with padding as a stencil

In a follow up PR I will change the Pool2D to use that generator and
also drop the 4-D inner constraint.

Note: For CPU kernels we prefer NHWC like formats where we can
vectorize in inner most, so stencil axes not being inner most could be a requirement for vectorization.
ALGORITHM_ORIG_REV_ID: 237b91192914ac6f2d171acbf19ececf3a9a8520