[Stdlib] Unify all uses of getting global variable from runtime (#24098)
We had scattered uses of KGEN_CompilerRT_GetGlobalOr and using it to
load and store the global variable. This unifies them into a single
place. It also unifies how we load functions in the case this is a
dylib.

GPU_ORIG_REV_ID: 4fe6f20248fdf225e2c82318661102b72440ddc9