[Kernels][GPU] Add GPU reduce ND kernel supporting fusion (#24825)
Currently only supports reduction over the inner dimension. Can
be extended to non-inner dim reduction in a followup.

Strategy is to assign each thread block to a row. The thread block
first does a partial accumulation of the row in register. Then
it does a parallel block reduction using warp shuffle to reduce to a
single element. By keeping the partial reduction within a block we
avoid allocating temporary memory. This strategy is suboptimal
if the row dim is much larger than the product of the other dims,
because each row only uses BLOCK_SIZE (=128) threads for the row
reduction.

Will be added to the kernel lib and plumbed through MOGG in a followup.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
GPU_TEST_ORIG_REV_ID: 03a8068aeef5a1b1e0ba090637afb6c0fbcc720b