[Kernels][GPU] Remove warp_id() because it is a footgun (#24829)
warp_id returns the id of the warp in the SM, not the id of the warp in
the thread block like you would expect. It is also volatile and can
change under the hood.

Removing since it is a footgun and source of confusion
https://stackoverflow.com/questions/44337309/whats-the-most-efficient-way-to-calculate-the-warp-id-lane-id-in-a-1-d-grid.

Co-authored-by: Ian Tramble <ian@modular.com>
GPU_ORIG_REV_ID: 10d09290ea0c84ecd7b6338888d9248c8da78431