[mojo-lib] Make testing functions raise on failure (#24720)
The current ergonomics of `assert_true`, `assert_false`, and friends to return a
Bool of whether the condition was met or not is not very friendly to
callers.  It forces callers to consume the return value and branch in order to
stop the execution on a given failure in a test.  For example, see the wrapped
`assert_true` and `assert_false` that was in `test_object.mojo`.

Change these `assert` functions to instead raise an `Error` so the ergonomics
of callers are improved to just say:

```
assert_true(some_condition, some_optional_message)
```

which will stop execution if `some_condition` is `False`.

Adjust the uses in tests now that the `assert` test functions can raise
an `Error`.

NN_ORIG_REV_ID: 1d7cd1d1f84ac99aa67efdd0f34e702715c3fdbe