[Kernels] Mogg <-> Kernel static shape passing for Matmul (#24892)
This changes the signature of the matmul in mogg to allow it to accept
static shapes. It then propagates those shape information to the
kernels. Note that there is some work to be done on partitioning to make
sure that we are not loosing information.

Future PRs will use this information to make sure that we partition the
CUDA launches as effectively as possible.

MOGG_ORIG_REV_ID: 859cd4142b45bd5e9c05876965c29b55e8cc025f