[Mojo] Align the mblack formatter options with the mojo format options (#25358)
This is long standing issue where the editor was formating code in a
certain way and the regular mblack was formating it in another way. We
just need to be consistent to avoid churn from the developer due to
minor formating differences. This chooses the options chosen by mojo
format which seem to be sensible.

Closes #23330

BUFFER_ORIG_REV_ID: 630f4ccae7af76ee33c34db28d433945f7610bab