[TF] Add support for custom ops (#25304)
This enables replacing TF ops with mojo code.
The mojo code just needs to be registered as a kernel with the same name as the TF op.
Custom ops take precedence over all other lowerings.

As part of the PR, the feature is exposed as a flag for mt and the tf-to-mo pass.

MOGG_ORIG_REV_ID: 4a23f56bc5d4b5c1036568a88f4cc8763260e84a