[Kernels][GPU] Propagate elementwise fn to matmul kernel (#25359)
Propagate the elementwise epilogue function to the GPU matmul kernel.
This allows us to lightup the fast matmul path in the model. There is
some things that I would like to improve upon this PR (especially in the
area of dispatch), but that can be done as a followup cleanup work

Closes [Internal link]

LINALG_ORIG_REV_ID: 16a217eccf3c32f6fa88d84489b00b88c3d83a86