[mojo-stdlib] Include sort in algorithm package (#25647)
This looks to be an oversight when it was originally added.  This makes it so
the `sort` symbols are included in the `algorithm` stdlib package like we do for
the functional and reduction files.  This only affects callers that did `import
algorithm` as the `sort` symbols would now be included.

ALGORITHM_ORIG_REV_ID: 89d026eb6803f5e8ade4d040a9356a63476b4257