[Kernels] Parallel GEMV implementation with a 20X speedup. (#26224)
Added support for an optionally parallel GEMV implementation.

The accompanying benchmark/test code runs the new gemv op with a 1024 x
1024 input matrix in both serial and parallel configurations. On a
MacBook Pro M3 Pro the serial version already saturates the bus with an
"effective bandwidth*" of ~150GB/s. The parallel version (on the six
performance cores) performs at 356GB/s with a ~2x speedup. With an input
matrix of 4096 x 4096 the speedup is ~5.4x, and with a 16384 x 16384
matrix ~5.8x.

On an intel based AWS node (c6i.4xlarge) the parallel speedup (with
respect to the previous serial implementation) is (with default number
of Runtime threads) ~4x for the 1024x1024 case, ~13x for the 4096 x 4096
case, and ~10x for a 16384 x 16384 matrix.

The algorithm hyper parameters are the thread group size, and the
vertical and horizontal blocking factors (col_unroll_factor and
row_block_size in the code). The optimal values for these parameters
will be different for various input data sizes and on various hardware
and OS configurations and should be optimized using Autotune.

* By "effective bandwidth" we mean the amount of data GEMV needs to
access (2 * rows * columns * sizeof(float)) divided by its latency.

LINALG_ORIG_REV_ID: 2efc4ec8b549a090c66061bc630140cbadab2265