[Stdlib] Do not unroll to decrease reg pressure on GPUs, NFC (#26413)
Do not unroll _get_nd_indices_from_flat_index on nvidia hardware to
reduce the number of registers that are live. This also switches to use
cheaper versions of div/mod because we know that the indecies are
positive.

ALGORITHM_ORIG_REV_ID: cbac2aeaa0c514706867343ca23df69bd09221d3